Finish replacing configure dialog with a drawer
