+++ enclave-manager/web/packages/app/src/emui/enclaves/components/CreateEnclave.tsx
@@ -2,16 +2,14 @@ import { KurtosisPackage } from "kurtosis-cloud-indexer-sdk";
 import { isDefined } from "kurtosis-ui-components";
 import { useCallback, useEffect, useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
-import { ConfigureEnclaveModal } from "./modals/ConfigureEnclaveModal";
-import { KURTOSIS_CREATE_ENCLAVE_URL_ARG } from "./modals/constants";
-import { ManualCreateEnclaveModal } from "./modals/ManualCreateEnclaveModal";
+import { KURTOSIS_CREATE_ENCLAVE_URL_ARG } from "./configuration/drawer/constants";
+import { CreateOrConfigureEnclaveDrawer } from "./configuration/drawer/CreateOrConfigureEnclaveDrawer";
 import { PreloadPackage } from "./PreloadPackage";
 
 export const CreateEnclave = () => {
   const navigate = useNavigate();
   const location = useLocation();
 
-  const [configureEnclaveOpen, setConfigureEnclaveOpen] = useState(false);
   const [kurtosisPackage, setKurtosisPackage] = useState<KurtosisPackage>();
   const [manualCreateEnclaveOpen, setManualCreateEnclaveOpen] = useState(false);
 
@@ -19,15 +17,9 @@ export const CreateEnclave = () => {
     setManualCreateEnclaveOpen(location.hash === `#${KURTOSIS_CREATE_ENCLAVE_URL_ARG}`);
   }, [location]);
 
-  const handleManualCreateEnclaveConfirmed = (kurtosisPackage: KurtosisPackage) => {
-    setKurtosisPackage(kurtosisPackage);
-    setManualCreateEnclaveOpen(false);
-    setConfigureEnclaveOpen(true);
-  };
-
   const handleOnPackageLoaded = useCallback((kurtosisPackage: KurtosisPackage) => {
     setKurtosisPackage(kurtosisPackage);
-    setConfigureEnclaveOpen(true);
+    setManualCreateEnclaveOpen(true);
   }, []);
 
   const handleCloseManualCreateEnclave = () => {
@@ -40,18 +32,11 @@ export const CreateEnclave = () => {
   return (
     <>
       <PreloadPackage onPackageLoaded={handleOnPackageLoaded} />
-      <ManualCreateEnclaveModal
+      <CreateOrConfigureEnclaveDrawer
         isOpen={manualCreateEnclaveOpen}
         onClose={handleCloseManualCreateEnclave}
-        onConfirm={handleManualCreateEnclaveConfirmed}
+        kurtosisPackage={kurtosisPackage}
       />
-      {isDefined(kurtosisPackage) && (
-        <ConfigureEnclaveModal
-          isOpen={configureEnclaveOpen}
-          onClose={() => setConfigureEnclaveOpen(false)}
-          kurtosisPackage={kurtosisPackage}
-        />
-      )}
     </>
   );
 };
