+++ enclave-manager/web/packages/app/src/emui/enclaves/components/configuration/drawer/DrawerExpandCollapseButton.tsx
@@ -1,0 +1,19 @@
+import { IconButton } from "@chakra-ui/react";
+import { BiArrowToLeft, BiArrowToRight } from "react-icons/bi";
+import { DrawerSizes } from "./types";
+
+type DrawerExpandCollapseButtonProps = {
+  drawerSize: DrawerSizes;
+  onClick: () => void;
+};
+export const DrawerExpandCollapseButton = ({ drawerSize, onClick }: DrawerExpandCollapseButtonProps) => {
+  return (
+    <IconButton
+      size={"sm"}
+      icon={drawerSize === "xl" ? <BiArrowToLeft /> : <BiArrowToRight />}
+      aria-label={"Expand/collapse"}
+      variant={"ghost"}
+      onClick={onClick}
+    />
+  );
+};
