+++ enclave-manager/web/packages/app/src/emui/enclaves/components/configuration/drawer/bodies/EnclaveConfigureBody.tsx
@@ -1,24 +1,16 @@
-import { InfoIcon, SmallCloseIcon } from "@chakra-ui/icons";
+import { InfoIcon } from "@chakra-ui/icons";
 import {
   Button,
   ButtonGroup,
   Card,
   CardBody,
-  Drawer,
   DrawerBody,
-  DrawerCloseButton,
-  DrawerContent,
   DrawerFooter,
   DrawerHeader,
-  DrawerOverlay,
   Flex,
   FormControl,
-  Icon,
-  IconButton,
-  Input,
-  InputGroup,
-  InputLeftElement,
-  InputRightElement,
+  Grid,
+  GridItem,
   TabPanel,
   TabPanels,
   Tabs,
@@ -31,216 +23,32 @@ import { ArgumentValueType, KurtosisPackage } from "kurtosis-cloud-indexer-sdk";
 import {
   assertDefined,
   CopyButton,
-  FindCommand,
   HoverLineTabList,
   isDefined,
   KurtosisAlert,
   KurtosisAlertError,
-  KurtosisPackageCardHorizontal,
   PackageLogo,
   readablePackageName,
   stringifyError,
-  useSavedPackages,
 } from "kurtosis-ui-components";
-import { ChangeEvent, useMemo, useRef, useState } from "react";
+import { useMemo, useRef, useState } from "react";
 import { ErrorBoundary } from "react-error-boundary";
 import { SubmitHandler } from "react-hook-form";
-import { BiArrowToLeft, BiArrowToRight } from "react-icons/bi";
-import { FiSearch, FiShare } from "react-icons/fi";
+import { FiShare } from "react-icons/fi";
 import { useNavigate } from "react-router-dom";
-import { useKurtosisClient } from "../../../../client/enclaveManager/KurtosisClientContext";
-import { CatalogContextProvider, useCatalogContext } from "../../../catalog/CatalogContext";
-import { useEnclavesContext } from "../../EnclavesContext";
-import { EnclaveFullInfo } from "../../types";
-import {
-  EnclaveConfigurationForm,
-  EnclaveConfigurationFormImperativeAttributes,
-} from "../configuration/EnclaveConfigurationForm";
-import { BooleanArgumentInput } from "../configuration/inputs/BooleanArgumentInput";
-import { StringArgumentInput } from "../configuration/inputs/StringArgumentInput";
-import { KurtosisArgumentFormControl } from "../configuration/KurtosisArgumentFormControl";
-import { KurtosisPackageArgumentInput } from "../configuration/KurtosisPackageArgumentInput";
-import { ConfigureEnclaveForm } from "../configuration/types";
-import { allowedEnclaveNamePattern, isEnclaveNameAllowed } from "../utils";
-import { KURTOSIS_PACKAGE_ID_URL_ARG, KURTOSIS_PACKAGE_PARAMS_URL_ARG } from "./constants";
-
-type DrawerSizes = "xl" | "full";
-
-type ManualCreateEnclaveModalProps = {
-  isOpen: boolean;
-  onClose: () => void;
-  onConfirm: (kurtosisPackage: KurtosisPackage) => void;
-};
-
-export const ManualCreateEnclaveModal = ({ isOpen, onClose, onConfirm }: ManualCreateEnclaveModalProps) => {
-  const [drawerSize, setDrawerSize] = useState<DrawerSizes>("xl");
-  const [kurtosisPackage, setKurtosisPackage] = useState<KurtosisPackage | null>(null);
-
-  const handleToggleDrawerSize = () => {
-    setDrawerSize((drawerSize) => (drawerSize === "xl" ? "full" : "xl"));
-  };
-
-  const handleClose = () => {
-    setKurtosisPackage(null);
-    onClose();
-  };
-
-  return (
-    <Drawer isOpen={isOpen} onClose={handleClose} size={drawerSize}>
-      <DrawerOverlay />
-      <DrawerContent>
-        <DrawerCloseButton />
-        <CatalogContextProvider>
-          {!isDefined(kurtosisPackage) && (
-            <PackageSelectBody
-              onPackageSelected={setKurtosisPackage}
-              onClose={handleClose}
-              drawerSize={drawerSize}
-              onDrawerSizeClick={handleToggleDrawerSize}
-            />
-          )}
-          {isDefined(kurtosisPackage) && (
-            <EnclaveConfigureBody
-              kurtosisPackage={kurtosisPackage}
-              onBackClicked={() => setKurtosisPackage(null)}
-              onClose={handleClose}
-              drawerSize={drawerSize}
-              onDrawerSizeClick={handleToggleDrawerSize}
-            />
-          )}
-        </CatalogContextProvider>
-      </DrawerContent>
-    </Drawer>
-  );
-};
-
-type DrawerExpandCollapseButtonProps = {
-  drawerSize: DrawerSizes;
-  onClick: () => void;
-};
-
-const DrawerExpandCollapseButton = ({ drawerSize, onClick }: DrawerExpandCollapseButtonProps) => {
-  return (
-    <IconButton
-      size={"sm"}
-      icon={drawerSize === "xl" ? <BiArrowToLeft /> : <BiArrowToRight />}
-      aria-label={"Expand/collapse"}
-      variant={"ghost"}
-      onClick={onClick}
-    />
-  );
-};
-
-type PackageSelectBodyProps = {
-  onPackageSelected: (kurtosisPackage: KurtosisPackage) => void;
-  onClose: () => void;
-  drawerSize: DrawerSizes;
-  onDrawerSizeClick: () => void;
-};
-
-const PackageSelectBody = ({ onPackageSelected, onClose, drawerSize, onDrawerSizeClick }: PackageSelectBodyProps) => {
-  const searchRef = useRef<HTMLInputElement>(null);
-  const [searchTerm, setSearchTerm] = useState("");
-
-  const { catalog } = useCatalogContext();
-
-  const searchResults = useMemo(
-    () =>
-      catalog.map((catalog) =>
-        catalog.packages.filter(
-          (kurtosisPackage) => kurtosisPackage.name.toLowerCase().indexOf(searchTerm.toLowerCase()) >= 0,
-        ),
-      ),
-    [catalog, searchTerm],
-  );
-
-  const { savedPackages, togglePackageSaved } = useSavedPackages();
-
-  const handleSearchTermChange = async (e: ChangeEvent<HTMLInputElement>) => {
-    setSearchTerm(e.target.value);
-  };
-
-  if (searchResults.isErr) {
-    return (
-      <DrawerBody>
-        <KurtosisAlert message={"Unable to load kurtosis packages"} details={searchResults.error} />
-      </DrawerBody>
-    );
-  }
-
-  return (
-    <>
-      <DrawerHeader display={"flex"} justifyContent={"space-between"} alignItems={"center"} width={"100%"}>
-        <DrawerExpandCollapseButton drawerSize={drawerSize} onClick={onDrawerSizeClick} />
-        <Text as={"span"}>Enclave Configuration</Text>
-        {/*Here to balance the space-between*/}
-        <Text />
-      </DrawerHeader>
-      <DrawerBody>
-        <InputGroup variant={"solid"} width={"100%"} color={"gray.150"}>
-          <InputLeftElement>
-            <Icon as={FiSearch} />
-          </InputLeftElement>
-          <Input
-            ref={searchRef}
-            value={searchTerm}
-            bgColor={"gray.850"}
-            onChange={handleSearchTermChange}
-            placeholder={"Search"}
-            autoFocus
-          />
-          <InputRightElement w={"unset"} mr={"8px"}>
-            {searchTerm.length > 0 ? (
-              <Button variant="ghost" size={"xs"} rightIcon={<SmallCloseIcon />} onClick={() => setSearchTerm("")}>
-                Clear
-              </Button>
-            ) : (
-              <FindCommand whiteSpace={"nowrap"} pr={"10px"} />
-            )}
-          </InputRightElement>
-        </InputGroup>
-        {(searchTerm.length > 0 || savedPackages.length === 0) && (
-          <Flex flexDirection={"column"} gap={"10px"}>
-            <Text fontWeight={"semibold"} pt={"16px"} pb={"6px"}>
-              {searchTerm.length === 0 ? "All Packages" : "Search Results"}
-            </Text>
-            {searchResults.value.map((kurtosisPackage) => (
-              <KurtosisPackageCardHorizontal
-                key={kurtosisPackage.name}
-                kurtosisPackage={kurtosisPackage}
-                onClick={() => onPackageSelected(kurtosisPackage)}
-              />
-            ))}
-          </Flex>
-        )}
-        {searchTerm.length === 0 && savedPackages.length > 0 && (
-          <Flex flexDirection={"column"} gap={"10px"}>
-            <Text fontWeight={"semibold"} pt={"16px"} pb={"6px"}>
-              Saved
-            </Text>
-            {savedPackages.map((kurtosisPackage) => (
-              <KurtosisPackageCardHorizontal
-                key={kurtosisPackage.name}
-                kurtosisPackage={kurtosisPackage}
-                onClick={() => onPackageSelected(kurtosisPackage)}
-              />
-            ))}
-          </Flex>
-        )}
-      </DrawerBody>
-      <DrawerFooter>
-        <Flex justifyContent={"flex-end"} gap={"12px"}>
-          <Button color={"gray.100"} onClick={onClose}>
-            Cancel
-          </Button>
-          <Button type={"submit"} colorScheme={"kurtosisGreen"}>
-            Configure
-          </Button>
-        </Flex>
-      </DrawerFooter>
-    </>
-  );
-};
+import { useKurtosisClient } from "../../../../../../client/enclaveManager/KurtosisClientContext";
+import { useEnclavesContext } from "../../../../EnclavesContext";
+import { EnclaveFullInfo } from "../../../../types";
+import { allowedEnclaveNamePattern, isEnclaveNameAllowed } from "../../../utils";
+import { EnclaveConfigurationForm, EnclaveConfigurationFormImperativeAttributes } from "../../EnclaveConfigurationForm";
+import { BooleanArgumentInput } from "../../inputs/BooleanArgumentInput";
+import { StringArgumentInput } from "../../inputs/StringArgumentInput";
+import { KurtosisArgumentFormControl } from "../../KurtosisArgumentFormControl";
+import { KurtosisPackageArgumentInput } from "../../KurtosisPackageArgumentInput";
+import { ConfigureEnclaveForm } from "../../types";
+import { KURTOSIS_PACKAGE_ID_URL_ARG, KURTOSIS_PACKAGE_PARAMS_URL_ARG } from "../constants";
+import { DrawerExpandCollapseButton } from "../DrawerExpandCollapseButton";
+import { DrawerSizes } from "../types";
 
 type EnclaveConfigureBodyProps = {
   kurtosisPackage: KurtosisPackage;
@@ -250,10 +58,8 @@ type EnclaveConfigureBodyProps = {
   onDrawerSizeClick: () => void;
   existingEnclave?: EnclaveFullInfo;
 };
-
 const tabs = ["Form", "YAML"] as const;
-
-const EnclaveConfigureBody = ({
+export const EnclaveConfigureBody = ({
   kurtosisPackage,
   onBackClicked,
   onClose,
@@ -431,6 +237,7 @@ const EnclaveConfigureBody = ({
 
     try {
       const logsIterator = await runStarlarkPackage(enclave, kurtosisPackage.name, submissionData);
+      onClose();
       navigator(`/enclave/${enclaveUUID}/logs`, { state: { logs: logsIterator } });
     } catch (error: any) {
       setError(stringifyError(error));
@@ -439,27 +246,30 @@ const EnclaveConfigureBody = ({
 
   return (
     <ErrorBoundary fallbackRender={KurtosisAlertError}>
-      <DrawerHeader display={"flex"} justifyContent={"space-between"} alignItems={"center"} width={"100%"}>
-        <ButtonGroup gap={"10px"}>
-          <DrawerExpandCollapseButton drawerSize={drawerSize} onClick={onDrawerSizeClick} />
-          <Tooltip shouldWrapChildren label={"Create a link that can be used to share this configuration."}>
-            <CopyButton
-              contentName={"url"}
-              valueToCopy={getLinkToCurrentConfig}
-              aria-label={"Copy link"}
-              isIconButton
-              color={"gray.100"}
-              icon={<FiShare />}
-              size={"sm"}
-            />
-          </Tooltip>
-        </ButtonGroup>
-        <Flex alignItems={"center"} gap={"8px"}>
-          <PackageLogo kurtosisPackage={kurtosisPackage} h={"24px"} display={"inline"} />
-          <Text as={"span"}>{readablePackageName(kurtosisPackage.name)}</Text>
-        </Flex>
-        {/*Here to balance the space-between*/}
-        <Text />
+      <DrawerHeader as={Grid} gridTemplateColumns={"1fr 1fr 1fr"}>
+        <GridItem>
+          <ButtonGroup gap={"10px"}>
+            <DrawerExpandCollapseButton drawerSize={drawerSize} onClick={onDrawerSizeClick} />
+            <Tooltip shouldWrapChildren label={"Create a link that can be used to share this configuration."}>
+              <CopyButton
+                contentName={"url"}
+                valueToCopy={getLinkToCurrentConfig}
+                aria-label={"Copy link"}
+                isIconButton
+                color={"gray.100"}
+                icon={<FiShare />}
+                size={"sm"}
+              />
+            </Tooltip>
+          </ButtonGroup>
+        </GridItem>
+        <GridItem>
+          <Flex alignItems={"center"} justifyContent={"center"} gap={"8px"}>
+            <PackageLogo kurtosisPackage={kurtosisPackage} h={"24px"} display={"inline"} />
+            <Text as={"span"}>{readablePackageName(kurtosisPackage.name)}</Text>
+          </Flex>
+        </GridItem>
+        <GridItem />
       </DrawerHeader>
       <EnclaveConfigurationForm
         ref={formRef}
@@ -529,7 +339,7 @@ const EnclaveConfigureBody = ({
           </Tabs>
         </DrawerBody>
         <DrawerFooter>
-          <Flex justifyContent={"space-between"} gap={"12px"}>
+          <Flex justifyContent={"space-between"} gap={"12px"} width={"100%"}>
             <Button color={"gray.100"} onClick={onBackClicked} isDisabled={isLoading}>
               Back
             </Button>
