+++ enclave-manager/web/packages/app/src/emui/enclaves/components/configuration/drawer/constants.ts
