+++ enclave-manager/web/packages/app/src/emui/enclaves/components/configuration/drawer/types.ts
@@ -1,0 +1,1 @@
+export type DrawerSizes = "xl" | "full";
